backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
mediapipe==0.10.7
pillow==10.0.1
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.4.2
fastapi==0.103.1
uvicorn==0.23.2 